Use raw SQL query and map to OrganisationDto
