+++ Services/CO.CDP.OrganisationInformation.Persistence/IOrganisationRepository.cs
@@ -31,6 +31,8 @@ public interface IOrganisationRepository : IDisposable
 
     public Task<IList<Organisation>> GetPaginated(PartyRole? role, PartyRole? pendingRole, string? searchText, int limit, int skip);
 
+    public Task<IList<DatabaseOrganisationRepository.OrganisationRawDto>> GetPaginatedRaw(PartyRole? role, PartyRole? pendingRole, string? searchText, int limit, int skip);
+
     public Task<int> GetTotalCount(PartyRole? role, PartyRole? pendingRole, string? searchText);
 
     public class OrganisationRepositoryException(string message, Exception? cause = null) : Exception(message, cause)
